requirements.txt
Django~=3.2.9
mysqlclient
pillow
slugify
django_jalali